requirements.txt
Flask==3.0.0
line-bot-sdk==3.13.0
gunicorn==20.1.0